Separate kebab and snake into their own files
