+++ index.d.ts
@@ -10,7 +10,9 @@ export {RequireAtLeastOne} from './source/require-at-least-one';
 export {RequireExactlyOne} from './source/require-exactly-one';
 export {PartialDeep} from './source/partial-deep';
 export {ReadonlyDeep} from './source/readonly-deep';
-export {DelimiterCase, KebabCase, SnakeCase} from './source/delimiter-case';
+export {DelimiterCase} from './source/delimiter-case';
+export {KebabCase} from './source/kebab-case';
+export {SnakeCase} from './source/snake-case';
 export {LiteralUnion} from './source/literal-union';
 export {Promisable} from './source/promisable';
 export {Opaque} from './source/opaque';
