+++ test-d/delimiter-case.ts
@@ -2,9 +2,7 @@ import {
 	WordSeparators,
 	UpperCaseChars,
 	SplitIncludingDelimiters,
-	DelimiterCase,
-	KebabCase,
-	SnakeCase
+	DelimiterCase
 } from '../source/delimiter-case';
 import {expectType, expectAssignable} from 'tsd';
 
@@ -34,49 +32,9 @@ expectType<'foo#bar'>(delimiterFromSnake);
 const noDelimiterFromMono: DelimiterCase<'foobar', '#'> = 'foobar';
 expectType<'foobar'>(noDelimiterFromMono);
 
-// KebabCase
-const kebabFromCamel: KebabCase<'fooBar'> = 'foo-bar';
-expectType<'foo-bar'>(kebabFromCamel);
-
-const kebabFromKebab: KebabCase<'foo-bar'> = 'foo-bar';
-expectType<'foo-bar'>(kebabFromKebab);
-
-const kebabFromSpace: KebabCase<'foo bar'> = 'foo-bar';
-expectType<'foo-bar'>(kebabFromSpace);
-
-const kebabFromSnake: KebabCase<'foo_bar'> = 'foo-bar';
-expectType<'foo-bar'>(kebabFromSnake);
-
-const noKebabFromMono: KebabCase<'foobar'> = 'foobar';
-expectType<'foobar'>(noKebabFromMono);
-
-// SnakeCase
-const snakeFromCamel: SnakeCase<'fooBar'> = 'foo_bar';
-expectType<'foo_bar'>(snakeFromCamel);
-
-const snakeFromKebab: SnakeCase<'foo-bar'> = 'foo_bar';
-expectType<'foo_bar'>(snakeFromKebab);
-
-const snakeFromSpace: SnakeCase<'foo bar'> = 'foo_bar';
-expectType<'foo_bar'>(snakeFromSpace);
-
-const snakeFromSnake: SnakeCase<'foo_bar'> = 'foo_bar';
-expectType<'foo_bar'>(snakeFromSnake);
-
-const noSnakeFromMono: SnakeCase<'foobar'> = 'foobar';
-expectType<'foobar'>(noSnakeFromMono);
-
 // Verifying example
-type KebabCasedProps<T> = {
-	[K in keyof T as DelimiterCase<K, '-'>]: T[K]
-};
-
-type AlternativeKebabCasedProps<T> = {
-	[K in keyof T as KebabCase<K>]: T[K]
-};
-
-type SnakeCasedProps<T> = {
-	[K in keyof T as SnakeCase<K>]: T[K]
+type OddCasedProps<T> = {
+	[K in keyof T as DelimiterCase<K, '#'>]: T[K]
 };
 
 interface CliOptions {
@@ -85,20 +43,8 @@ interface CliOptions {
 	foo: number;
 }
 
-expectAssignable<KebabCasedProps<CliOptions>>({
-	'dry-run': true,
-	'include-file': 'bar.js',
-	foo: 123
-});
-
-expectAssignable<AlternativeKebabCasedProps<CliOptions>>({
-	'dry-run': true,
-	'include-file': 'bar.js',
-	foo: 123
-});
-
-expectAssignable<SnakeCasedProps<CliOptions>>({
-	dry_run: true,
-	include_file: 'bar.js',
+expectAssignable<OddCasedProps<CliOptions>>({
+	'dry#run': true,
+	'include#file': 'bar.js',
 	foo: 123
 });
