+++ test-d/snake-case.ts
@@ -1,0 +1,19 @@
+import {
+	SnakeCase
+} from '../source/snake-case';
+import {expectType} from 'tsd';
+
+const snakeFromCamel: SnakeCase<'fooBar'> = 'foo_bar';
+expectType<'foo_bar'>(snakeFromCamel);
+
+const snakeFromKebab: SnakeCase<'foo-bar'> = 'foo_bar';
+expectType<'foo_bar'>(snakeFromKebab);
+
+const snakeFromSpace: SnakeCase<'foo bar'> = 'foo_bar';
+expectType<'foo_bar'>(snakeFromSpace);
+
+const snakeFromSnake: SnakeCase<'foo_bar'> = 'foo_bar';
+expectType<'foo_bar'>(snakeFromSnake);
+
+const noSnakeFromMono: SnakeCase<'foobar'> = 'foobar';
+expectType<'foobar'>(noSnakeFromMono);
